architect POC/requirements.txt
pydantic
httpx
beautifulsoup4
python-dotenv
google-genai
